+++ src/libromdata/Console/DreamcastSave.cpp
@@ -1490,6 +1490,54 @@ int DreamcastSave::loadFieldData(void)
 }
 
 /**
+ * Load metadata properties.
+ * Called by RomData::metaData() if the field data hasn't been loaded yet.
+ * @return Number of metadata properties read on success; negative POSIX error code on error.
+ */
+int DreamcastSave::loadMetaData(void)
+{
+	RP_D(DreamcastSave);
+	if (d->metaData != nullptr) {
+		// Metadata *has* been loaded...
+		return 0;
+	} else if (!d->file) {
+		// File isn't open.
+		return -EBADF;
+	} else if (!d->isValid || (int)d->saveType < 0) {
+		// Unknown save type.
+		return -EIO;
+	}
+
+	// Create the metadata object.
+	d->metaData = new RomMetaData();
+	d->metaData->reserve(2);	// Maximum of 4 metadata properties.
+
+	// TODO: More metadata properties?
+
+	// Title.
+	if (d->loaded_headers & DreamcastSavePrivate::DC_HAVE_VMS) {
+		d->metaData->addMetaData_string(Property::Title,
+				cp1252_sjis_to_utf8(
+					d->vms_header.dc_description, sizeof(d->vms_header.dc_description)),
+					RomFields::STRF_TRIM_END);
+	} else if (d->loaded_headers & DreamcastSavePrivate::DC_HAVE_VMI) {
+		d->metaData->addMetaData_string(Property::Title,
+			cp1252_sjis_to_utf8(
+				d->vmi_header.description, sizeof(d->vmi_header.description)),
+				RomFields::STRF_TRIM_END);
+	}
+
+	// Creation time.
+	if (d->loaded_headers & DreamcastSavePrivate::DC_HAVE_DIR_ENTRY) {
+		// TODO: Dreamcast doesn't support timezones.
+		d->metaData->addMetaData_timestamp(Property::CreationDate, d->ctime);
+	}
+
+	// Finished reading the metadata.
+	return static_cast<int>(d->metaData->count());
+}
+
+/**
  * Load an internal image.
  * Called by RomData::image().
  * @param imageType	[in] Image type to load.